Update config.py
The original text in the config file has comments saying `# ex: BLUEPRINTS = ['blog']  # where app is a Blueprint instance`.

Nothing on the line or in the lines nearby makes a reference to `app` - should the term `app` be changed to `blog`? Or am I misunderstanding the meaning of the comment? If I am misunderstanding, what could be done to make the comment more clear?